Fixes
* fixed incremental backup
* fixed s3 storage files deletion
* fixed working with no existing dirs on ftp
* binary files provides in tar archive with readme and license files
